requirements.txt
MacFSEvents==0.8.4
